+++ lightwallet-client-lib/src/main/java/co/electriccoin/lightwallet/client/internal/LightWalletClientImpl.kt
@@ -22,10 +22,14 @@ import io.grpc.Channel
 import io.grpc.ConnectivityState
 import io.grpc.ManagedChannel
 import io.grpc.StatusException
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.launch
 import java.util.concurrent.TimeUnit
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
@@ -47,6 +51,11 @@ internal class LightWalletClientImpl(
     private val streamingRequestTimeout: Duration = 90.seconds
 ) : LightWalletClient {
     private var channel = channelFactory.newChannel(lightWalletEndpoint)
+
+    // These make the LightWalletClientImpl not thread safe. In the long-term, we should
+    // consider making it thread safe.
+    private var stateCount = 0
+    private var state: ConnectivityState? = null
 
     override fun getBlockRange(heightRange: ClosedRange<BlockHeightUnsafe>): Flow<Response<CompactBlockUnsafe>> {
         require(!heightRange.isEmpty()) {
@@ -260,6 +269,30 @@ internal class LightWalletClientImpl(
             GrpcStatusResolver.resolveFailureFromStatus(e)
         }
 
+    override fun observeMempool(): Flow<Response<RawTransactionUnsafe>> =
+        channelFlow {
+            launch {
+                while (true) {
+                    requireChannel()
+                        .createStub(streamingRequestTimeout)
+                        .getMempoolStream(Service.Empty.newBuilder().build())
+                        .map {
+                            Response.Success(RawTransactionUnsafe.new(it))
+                        }.catch {
+                            // do nothing
+                        }.collect {
+                            send(it)
+                        }
+
+                    delay(30.seconds)
+                }
+            }
+
+            awaitClose {
+                // do nothing
+            }
+        }
+
     private fun shutdown() {
         channel.shutdown()
     }
@@ -268,11 +301,6 @@ internal class LightWalletClientImpl(
         channel.shutdown()
         channel = channelFactory.newChannel(lightWalletEndpoint)
     }
-
-    // These make the LightWalletClientImpl not thread safe. In the long-term, we should
-    // consider making it thread safe.
-    private var stateCount = 0
-    private var state: ConnectivityState? = null
 
     private fun requireChannel(): ManagedChannel {
         state =