+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Mempool is now being observed in order to store the transactions locally
+
 ### Fixed
 - Filtering transactions by memo now returns more than one item