+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `RustBackend.decryptAndStoreTransaction` now returns transaction ID
+
 ### Fixed
 - Filtering transactions by memo now returns more than one item